requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
functions-framework==3.5.0
urllib3==1.26.18
certifi==2024.2.2
google-cloud-secret-manager==2.20.0
